Moved bot_thread to another file.
