+++ source/botbot.py
@@ -1,14 +1,20 @@
-from websocket import create_connection, WebSocketConnectionClosedException
+#Standard library modules
 import json
 import time
 import sys
 import os
 import random
 import threading
+
 import re
 
+#Additional modules
+from websocket import create_connection, WebSocketConnectionClosedException
+
+#Project modules
 import logger
 import parser
+import bot_thread as bt
 
 log = logger.Logger()
 
@@ -20,7 +26,7 @@ room_name = 'testing'
 nickname = 'BotBot'
 snapshot_dir = 'snapshots'
 
-bots = []
+bots = bt.bots
 
 spam_threshold_messages = 10
 spam_threshold_time = 5
@@ -35,178 +41,10 @@ if __name__ == '__main__':
         room_name = sys.argv[1]
 
 web_socket_url = 'wss://euphoria.io/room/{}/ws'.format(room_name)
-log.log('Connecting to ' + web_socket_url + ' as ' + nickname + '...')
+log.write('Connecting to ' + web_socket_url + ' as ' + nickname + '...')
 ws = create_connection(web_socket_url)
 mid = 0
 agent_id = None
-
-class bot_thread (threading.Thread):
-    def __init__(self, thread_nickname, data, thread_room_name, creator):
-        threading.Thread.__init__(self)
-        self.paused = False
-        self.finished = False
-        self.last_times = []
-        self._kill = threading.Event()
-        self.data = data
-        self.mid = 0
-        self.room_name = thread_room_name
-        self.agent_id = None
-        self.web_socket_url = 'wss://euphoria.io/room/{}/ws'.format(self.room_name)
-        self.nickname = thread_nickname
-        self.creator = creator
-        self.help_text = '@' + self.nickname + ' is a bot created by \"' + creator + '\" using @' + nickname + '.\n\n@' + self.nickname + ' responds to !ping, !help @' + self.nickname + ', and the following regexes:\n' + '\n'.join(self.data.get_regexes()) + '\n\nTo pause this bot, use the command !pause @' + self.nickname + '.\nTo kill this bot, use the command !kill @' + self.nickname + '.'
-        self.pause_text = ''
-        log.log('Connecting to ' + self.web_socket_url + '...')
-        self.ws = create_connection(self.web_socket_url)
-        log.log('[' + self.nickname + '] Connected!')
-        self.thread_send_nick()
-    def run(self):
-        while not self.is_dead():
-            try:
-                if self.is_dead():
-                    self.ws.close()
-                    return
-                data = self.ws.recv()
-                if self.is_dead():
-                    self.ws.close()
-                    return
-            except WebSocketConnectionClosedException:
-                if self.is_dead():
-                    return
-                log.log('[' + self.nickname + '] Disconnected. Attempting to reconnect...')
-                self.ws = create_connection(web_socket_url)
-                self.thread_send_nick()
-                log.log('[' + self.nickname + '] Reconnected!')
-                data = self.ws.recv()
-                if self.is_dead():
-                    self.ws.close()
-                    return
-            data = json.loads(data)
-            if data['type'] == 'ping-event':
-                self.thread_send_ping()
-            if data['type'] == 'nick-reply':
-                self.agent_id = data['data']['id']
-            if data['type'] == 'send-event':
-                content = data['data']['content']
-                parent = data['data']['parent']
-                this_message = data['data']['id']
-                sender = data['data']['sender']['name']
-                send_time = data['data']['time']
-                sender_agent_id = data['data']['sender']['id']
-                self.recv_message(content, parent, this_message, sender, send_time, sender_agent_id, self.room_name)
-        self.finished = True
-
-    def recv_message(self, content='', parent=None, this_message=None, sender='', send_time=0, sender_agent_id='', room_name=''):
-        if sender_agent_id == agent_id:
-            return
-        if (len(content) == 7 + len(self.nickname) or (len(content) >= 7 + len(self.nickname) and content[7+len(self.nickname)] == ' ')) and content[0:7+len(self.nickname)].lower() == ('!kill @' + self.nickname).lower():
-            for bot in bots:
-                if sender_agent_id == bot.agent_id:
-                    return
-            self.thread_send_message('/me is now exiting.', this_message)
-            self.kill()
-        elif self.paused and ((len(content) == 10 + len(self.nickname) or (len(content) >= 10 + len(self.nickname) and content[10+len(self.nickname)] == ' ')) and content[0:10+len(self.nickname)].lower() == ('!restore @' + self.nickname).lower()):
-            for bot in bots:
-                if sender_agent_id == bot.agent_id:
-                    return
-            self.thread_send_message('/me is now restored.', this_message)
-            self.paused = False
-            self.pause_text = ''
-        elif (len(content) == 8 + len(self.nickname) or (len(content) >= 8 + len(self.nickname) and content[8+len(self.nickname)] == ' ')) and content[0:8+len(self.nickname)].lower() == ('!pause @' + self.nickname).lower():
-            for bot in bots:
-                if sender_agent_id == bot.agent_id:
-                    return
-            if self.paused:
-                self.thread_send_message('/me is already paused.', this_message)
-                self.thread_send_message('To restore this bot, type \"!restore @' + self.nickname + '\", or to kill this bot, type \"!kill @' + self.nickname + '\".', this_message)
-            else:
-                self.paused = True
-                self.pause_text = '/me has been paused by \"' + sender + '\".'
-                self.thread_send_message(self.pause_text, this_message)
-                self.thread_send_message('To restore this bot, type \"!restore @' + self.nickname + '\", or to kill this bot, type \"!kill @' + self.nickname + '\".', this_message)
-        elif self.paused and ((len(content) == 7 + len(self.nickname) or (len(content) >= 7 + len(self.nickname) and content[7+len(self.nickname)] == ' ')) and content[0:7+len(self.nickname)].lower() == ('!help @' + self.nickname).lower()):
-            self.thread_send_message(self.pause_text, this_message)
-            self.thread_send_message('To restore this bot, type \"!restore @' + self.nickname + '\", or to kill this bot, type \"!kill @' + self.nickname + '\".', this_message)
-        elif self.paused:
-            return
-        elif len(self.data.get_messages(content, sender)) > 0:
-            self.last_times.append(time.time())
-            while len(self.last_times) > spam_threshold_messages:
-                del self.last_times[0]
-            if not self.paused and len(self.last_times) == spam_threshold_messages:
-                if self.last_times[-1] - self.last_times[0] <= spam_threshold_time:
-                    ## Spam detected!
-                    self.paused = True
-                    self.pause_text = '/me has been temporarily halted due to a possible spam attack being generated by this bot.'
-                    self.thread_send_message(self.pause_text)
-                    self.thread_send_message('To restore this bot, type \"!restore @' + self.nickname + '\", or to kill this bot, type \"!kill @' + self.nickname + '\".')
-                    return
-            while len(self.last_times) > spam_threshold_messages - 1:
-                del self.last_times[0]
-            for x in self.data.get_messages(content, sender):
-                message = x.replace('(sender)', sender).replace('(@sender)', '@' + sender.replace(' ', '')).replace('(room)', room_name)
-                if re.match('^!ping\\b', message, re.IGNORECASE):
-                    continue
-                match = re.match('!to\\s+@(\\S+)\\s+&(\\S+)\\s+([\\s\\S]*)', message, re.IGNORECASE)
-                if match:
-                    for bot in bots:
-                        if bot.nickname.lower() == match.group(1).lower() and bot.room_name.lower() == match.group(2).lower():
-                            bot.recv_message(match.group(3), None, None, sender, send_time, sender_agent_id, room_name)
-                    continue
-                match = re.match('!to\\s+@(\\S+)\\s+([\\s\\S]*)', message, re.IGNORECASE)
-                if match:
-                    for bot in bots:
-                        if bot.nickname.lower() == match.group(1).lower():
-                            bot.recv_message(match.group(2), None, None, sender, send_time, sender_agent_id, room_name)
-                    continue
-                self.thread_send_message(message, this_message)
-        elif len(content) > 0 and content[0] == '!':
-            if len(content) >= 5 and content[1:5].lower() == 'ping':
-                self.thread_send_message('Pong!', this_message)
-                log.log('[' + self.nickname + '] Ponged to a ping from \"' + sender + '\".')
-            elif (len(content) == 7 + len(self.nickname) or (len(content) >= 7 + len(self.nickname) and content[7+len(self.nickname)] == ' ')) and content[1:7+len(self.nickname)].lower() == ('help @' + self.nickname).lower():
-                self.thread_send_message(self.help_text, this_message)
-                log.log('[' + self.nickname + '] Sent help text to \"' + sender + '\".')
-
-    def thread_send(self, message):
-        try:
-            self.ws.send(message)
-        except WebSocketConnectionClosedException:
-            self.ws = create_connection(self.web_socket_url)
-            self.ws.send(message)
-    def thread_send_ping(self):
-        reply = {'type':'ping-reply','data':{'time':int(time.time())},'id':str(self.mid)}
-        reply = json.dumps(reply)
-        self.thread_send(reply)
-        self.mid += 1
-    def thread_send_message(self, message, parent=None):
-        message = {'type':'send','data':{'content':message,'parent':parent},'id':str(self.mid)}
-        message = json.dumps(message)
-        self.thread_send(message)
-        self.mid += 1
-    def thread_send_nick(self, nick=None):
-        if nick == None:
-            nick = self.nickname
-        message = {'type':'nick','data':{'name':nick},'id':str(self.mid)}
-        message = json.dumps(message)
-        self.thread_send(message)
-        self.mid += 1
-    def kill(self):
-        self._kill.set()
-        self.ws.close()
-        try:
-            bots.remove(self)
-        except ValueError:
-            pass
-        log.log('[' + self.nickname + '] Exiting.')
-    def announce_and_kill(self, parent=None):
-        try:
-            self.thread_send_message('/me is now exiting.', parent)
-        except WebSocketConnectionClosedException:
-            pass
-        self.kill()
-    def is_dead(self):
-        return self._kill.isSet()
 
 def send(message):
     global ws
@@ -277,7 +115,7 @@ def create_snapshot(this_message=None, sender='(System)'):
         file.close()
         send_message('To load this snapshot later, type \"!load @' + nickname + ' ' + filename + '\".', this_message)
         send_message('Snapshot summary:\n' + '\n'.join(bot_names), this_message)
-        log.log('Created snapshot and sent to \"' + sender + '\".')
+        log.write('Created snapshot and sent to \"' + sender + '\".')
         return filepath
     except:
         send_message('Failed to create snapshot.')
@@ -298,7 +136,7 @@ def load_snapshot(filename, this_message=None, sender='(system)'):
         packed_bots = json.load(file)
         file.close()
         for packed_bot in packed_bots:
-            bot = bot_thread(packed_bot['nickname'][:36], parser.Parser(packed_bot['data']), packed_bot['room'], packed_bot['creator'])
+            bot = bt.bot_thread(packed_bot['nickname'][:36], parser.Parser(packed_bot['data']), packed_bot['room'], packed_bot['creator'])
             bots.append(bot)
             try:
                 bot.paused = packed_bot['paused']
@@ -310,10 +148,10 @@ def load_snapshot(filename, this_message=None, sender='(system)'):
                 pass
             bot.start()
         send_message('Successfully loaded snapshot.', this_message)
-        log.log('Loaded snapshot at \"' + filename + '\" from \"' + sender + '\".')
+        log.write('Loaded snapshot at \"' + filename + '\" from \"' + sender + '\".')
     except:
         send_message('Failed to load snapshot.', this_message)
-        log.log('Failed to load snapshot at \"' + filename + '\" from \"' + sender + '\".')
+        log.write('Failed to load snapshot at \"' + filename + '\" from \"' + sender + '\".')
 
 def killall_and_load_snapshot(filename, this_message=None, sender='(system)'):
     global bots
@@ -335,8 +173,8 @@ def killall_and_load_snapshot(filename, this_message=None, sender='(system)'):
 
 send_nick()
 send_message('/me Hello, world!')
-log.log('Connected!')
-log.log('Attempting to load most recent snapshot...')
+log.write('Connected!')
+log.write('Attempting to load most recent snapshot...')
 latest_snapshot_filename = None
 latest_snapshot = None
 snapshot_matching_regex = re.compile('^(\\d+)-(\\d+)-(\\d+)_(\\d\\d)(\\d\\d)(\\d\\d)(?:_(\\d+))?\.json$', re.IGNORECASE)
@@ -397,10 +235,10 @@ while True:
     try:
         data = ws.recv()
     except WebSocketConnectionClosedException:
-        log.log('Disconnected. Attempting to reconnect...')
+        log.write('Disconnected. Attempting to reconnect...')
         ws = create_connection(web_socket_url)
         send_nick()
-        log.log(' Reconnected!')
+        log.write(' Reconnected!')
         data = ws.recv()
     data = json.loads(data)
     if data['type'] == 'ping-event':
@@ -418,7 +256,7 @@ while True:
         if len(content) > 0 and content[0] == '!':
             if content[1:].lower() == 'ping':
                 send_message('Pong!', this_message)
-                log.log('Ponged to a ping from \"' + sender + '\".')
+                log.write('Ponged to a ping from \"' + sender + '\".')
             elif (len(content) == 7 + len(nickname) or (len(content) >= 7 + len(nickname) and content[7+len(nickname)] == ' ')) and content[1:7+len(nickname)].lower() == ('list @' + nickname).lower():
                 dont_respond = False
                 for bot in bots:
@@ -447,7 +285,7 @@ while True:
                     send_message('No bots created with @' + nickname + ' are running.', this_message)
                 else:
                     send_message('Currently running bots created with @' + nickname + ':\n' + '\n'.join(bot_names), this_message)
-                log.log('Listed bots by request from \"' + sender + '\".')
+                log.write('Listed bots by request from \"' + sender + '\".')
             elif (len(content) == 7 + len(nickname) or (len(content) >= 7 + len(nickname) and content[7+len(nickname)] == ' ')) and content[1:7+len(nickname)].lower() == ('help @' + nickname).lower():
                 dont_respond = False
                 for bot in bots:
@@ -457,7 +295,7 @@ while True:
                 if dont_respond:
                     continue
                 send_message(help_text, this_message)
-                log.log('Sent help text to \"' + sender + '\".')
+                log.write('Sent help text to \"' + sender + '\".')
             elif (len(content) == 10 + len(nickname) or (len(content) >= 10 + len(nickname) and content[10+len(nickname)] == ' ')) and content[1:10+len(nickname)].lower() == ('killall @' + nickname).lower():
                 dont_respond = False
                 for bot in bots:
@@ -469,7 +307,7 @@ while True:
                 send_message('Killing all bots...', this_message)
                 while len(bots) > 0:
                     bots.pop().announce_and_kill(this_message)
-                log.log('Killed all bots by request from \"' + sender + '\".')
+                log.write('Killed all bots by request from \"' + sender + '\".')
             elif len(content) > 13 and content[1:12].lower() == 'createbot @':
                 dont_respond = False
                 for bot in bots:
@@ -484,11 +322,11 @@ while True:
                 bot_nickname = parse_tree[1][1:][:36]
                 try:
                     bot_data = parser.Parser(parse_tree[2])
-                    bot = bot_thread(bot_nickname, bot_data, room_name, sender)
+                    bot = bt.bot_thread(bot_nickname, bot_data, room_name, sender, log)
                     bots.append(bot)
                     bot.start()
                     send_message('Created @' + bot_nickname + '.', this_message)
-                    log.log('Created @' + bot_nickname + ' by request from \"' + sender + '\".')
+                    log.write('Created @' + bot_nickname + ' by request from \"' + sender + '\".')
                 except:
                     send_message('Failed to create @' + bot_nickname + '. Is your code valid?', this_message)
                     send_message('Error details:\n' + str(sys.exc_info()[0]) + ': ' + str(sys.exc_info()[1]), this_message)
@@ -508,11 +346,11 @@ while True:
                 bot_nickname = parse_tree[2][1:][:36]
                 try:
                     bot_data = parser.Parser(parse_tree[3])
-                    bot = bot_thread(bot_nickname, bot_data, parse_tree[1][1:].lower(), sender)
+                    bot = bt.bot_thread(bot_nickname, bot_data, parse_tree[1][1:].lower(), sender)
                     bots.append(bot)
                     bot.start()
                     send_message('Created @' + bot_nickname + ' in ' + parse_tree[1].lower() + '.', this_message)
-                    log.log('Created @' + bot_nickname + ' by request from \"' + sender + '\" in &' + parse_tree[1][1:].lower() + '.')
+                    log.write('Created @' + bot_nickname + ' by request from \"' + sender + '\" in &' + parse_tree[1][1:].lower() + '.')
                 except:
                     send_message('Failed to create @' + bot_nickname + ' in ' + parse_tree[1].lower() + '. Is your code valid?', this_message)
                     send_message('Error details:\n' + str(sys.exc_info()[0]) + ': ' + str(sys.exc_info()[1]), this_message)
@@ -531,11 +369,11 @@ while True:
                 bot_nickname = parse_tree[1][1:][:36]
                 try:
                     bot_data = parser.Parser(parse_tree[2])
-                    bot = bot_thread(bot_nickname, bot_data, room_name, sender)
+                    bot = bt.bot_thread(bot_nickname, bot_data, room_name, sender, log)
                     bots.append(bot)
                     bot.start()
                     send_message('Created @' + bot_nickname + '.', this_message)
-                    log.log('Created @' + bot_nickname + ' by request from \"' + sender + '\".')
+                    log.write('Created @' + bot_nickname + ' by request from \"' + sender + '\".')
                 except:
                     send_message('Failed to create @' + bot_nickname + '. Is your code valid?', this_message)
                     send_message('Error details:\n' + str(sys.exc_info()[0]) + ': ' + str(sys.exc_info()[1]), this_message)
@@ -556,11 +394,11 @@ while True:
                 bot_nickname = parse_tree[2][1:][:36]
                 try:
                     bot_data = parser.Parser(parse_tree[3])
-                    bot = bot_thread(bot_nickname, bot_data, parse_tree[1][1:].lower(), sender)
+                    bot = bt.bot_thread(bot_nickname, bot_data, parse_tree[1][1:].lower(), sender, log)
                     bots.append(bot)
                     bot.start()
                     send_message('Created @' + bot_nickname + ' in ' + parse_tree[1].lower() + '.', this_message)
-                    log.log('Created @' + bot_nickname + ' by request from \"' + sender + '\" in &' + parse_tree[1][1:].lower() + '.')
+                    log.write('Created @' + bot_nickname + ' by request from \"' + sender + '\" in &' + parse_tree[1][1:].lower() + '.')
                 except:
                     send_message('Failed to create @' + bot_nickname + ' in ' + parse_tree[1].lower() + '. Is your code valid?', this_message)
                     send_message('Error details:\n' + str(sys.exc_info()[0]) + ': ' + str(sys.exc_info()[1]), this_message)
@@ -586,11 +424,11 @@ while True:
                 elif len(desired_bots) == 1:
                     try:
                         bot_data = parser.Parser(desired_bots[0].data.parse_string)
-                        bot = bot_thread(bot_nickname, bot_data, parse_tree[1][1:].lower(), desired_bots[0].creator)
+                        bot = bt.bot_thread(bot_nickname, bot_data, parse_tree[1][1:].lower(), desired_bots[0].creator, log)
                         bots.append(bot)
                         bot.start()
                         send_message('Created @' + bot_nickname + ' in ' + parse_tree[1].lower() + '.', this_message)
-                        log.log('Copied @' + bot_nickname + ' by request from \"' + sender + '\" to &' + parse_tree[1][1:].lower() + '.')
+                        log.write('Copied @' + bot_nickname + ' by request from \"' + sender + '\" to &' + parse_tree[1][1:].lower() + '.')
                     except:
                         send_message('Failed to create @' + bot_nickname + ' in ' + parse_tree[1].lower() + '. Is your code valid?', this_message)
                         send_message('Error details:\n' + str(sys.exc_info()[0]) + ': ' + str(sys.exc_info()[1]), this_message)
@@ -599,11 +437,11 @@ while True:
                         desired_bots = [desired_bots[int(parse_tree[3]) - 1]]
                         try:
                             bot_data = parser.Parser(desired_bots[0].data.parse_string)
-                            bot = bot_thread(bot_nickname, bot_data, parse_tree[1][1:].lower(), desired_bots[0].creator)
+                            bot = bt.bot_thread(bot_nickname, bot_data, parse_tree[1][1:].lower(), desired_bots[0].creator, log)
                             bots.append(bot)
                             bot.start()
                             send_message('Created @' + bot_nickname + ' in ' + parse_tree[1].lower() + '.', this_message)
-                            log.log('Copied @' + bot_nickname + ' by request from \"' + sender + '\" to &' + parse_tree[1][1:].lower() + '.')
+                            log.write('Copied @' + bot_nickname + ' by request from \"' + sender + '\" to &' + parse_tree[1][1:].lower() + '.')
                         except:
                             send_message('Failed to create @' + bot_nickname + ' in ' + parse_tree[1].lower() + '. Is your code valid?', this_message)
                             send_message('Error details:\n' + str(sys.exc_info()[0]) + ': ' + str(sys.exc_info()[1]), this_message)
